# Conflicts:
#	Examples/Benchmarks/Benchmarks.xcworkspace/xcshareddata/swiftpm/Package.resolved
#	Project.swift
#	Sources/EmbraceCommonInternal/Storage/Model/EmbraceSpan.swift
#	Sources/EmbraceConfigInternal/EmbraceConfig.swift
#	Sources/EmbraceConfigInternal/EmbraceConfigurable/RemoteConfig.swift
#	Sources/EmbraceConfiguration/EmbraceConfigurable.swift
#	Sources/EmbraceConfiguration/EmbraceConfigurable/DefaultConfig.swift
#	Sources/EmbraceConfiguration/SpanEventsLimits.swift
#	Sources/EmbraceCore/Capture/MetricKit/MetricKitCrashCaptureService+Options.swift
#	Sources/EmbraceCore/Capture/MetricKit/MetricKitHangCaptureService+Options.swift
#	Sources/EmbraceCore/Embrace.swift
#	Sources/EmbraceCore/Internal/Embrace+EmbraceMetricKitStateProvider.swift
#	Sources/EmbraceCore/Internal/Events/SpanEventsLimiter.swift
#	Sources/EmbraceCore/Internal/Logs/EmbraceLogAttributesBuilder.swift
#	Sources/EmbraceCore/Internal/Logs/Exporter/StorageEmbraceLogExporter.swift
#	Sources/EmbraceCore/Internal/ResourceStorageExporter/ResourceStorageExporter.swift
#	Sources/EmbraceCore/Internal/Tracing/EmbraceSpanProcessor+Setup.swift
#	Sources/EmbraceCore/Options/Embrace+Endpoints.swift
#	Sources/EmbraceCore/Options/Embrace+Options.swift
#	Sources/EmbraceCore/Payload/Builders/LogPayloadBuilder.swift
#	Sources/EmbraceCore/Payload/Builders/SpansPayloadBuilder.swift
#	Sources/EmbraceCore/Payload/Spans/SpanEventPayload.swift
#	Sources/EmbraceCore/Payload/Utils/PayloadUtils.swift
#	Sources/EmbraceCore/Public/Embrace+OTel.swift
#	Sources/EmbraceCore/Public/Metadata/MetadataHandler+Personas.swift
#	Sources/EmbraceCore/Public/OpenTelemetryProcessor.swift
#	Sources/EmbraceCore/Session/DataRecovery/UnsentDataHandler.swift
#	Sources/EmbraceCore/Session/SessionController.swift
#	Sources/EmbraceCore/Session/SessionSpanUtils.swift
#	Sources/EmbraceOTelInternal/DefaultEmbraceLogSharedState.swift
#	Sources/EmbraceSemantics/Log/LogSemantics+MetricKitMetrics.swift
#	Sources/EmbraceStorageInternal/EmbraceStorage.swift
#	Sources/EmbraceStorageInternal/Operations/EmbraceStorage+Span.swift
#	Sources/EmbraceStorageInternal/Records/LogRecord.swift
#	Sources/EmbraceStorageInternal/Records/SpanEventRecord.swift
#	Sources/EmbraceStorageInternal/Records/SpanRecord.swift
#	Tests/EmbraceConfigInternalTests/EmbraceConfigurable/RemoteConfigTests.swift
#	Tests/EmbraceCoreTests/Capture/WebView/WebViewCaptureServiceTests.swift
#	Tests/EmbraceCoreTests/Internal/EmbraceSpanProcessor+StorageTests.swift
#	Tests/EmbraceCoreTests/Internal/EmbraceSpanProcessorExporterTests.swift
#	Tests/EmbraceCoreTests/Internal/Events/SpanEventsLimiterTests.swift
#	Tests/EmbraceCoreTests/Internal/Logs/Exporter/StorageEmbraceLogExporterTests.swift
#	Tests/EmbraceCoreTests/Payload/SpansPayloadBuilderTests.swift
#	Tests/EmbraceCoreTests/Session/SessionSpanUtilsTests.swift
#	Tests/EmbraceCoreTests/Utils/W3C/W3C+TraceParentTests.swift
#	Tests/EmbraceOTelInternalTests/Trace/Tracer/Span/Processor/EmbraceSpanProcessorTests.swift
#	Tests/EmbraceStorageInternalTests/Records/SpanRecord/EmbraceStorage+SpanTests.swift
#	Tests/EmbraceStorageInternalTests/SpanRecordTests.swift
#	Tests/TestSupport/EditableConfig.swift
#	Tests/TestSupport/Mocks/MockEmbraceConfigurable.swift
#	bin/templates/EmbraceIO.podspec.tpl

+++ Sources/EmbraceUploadInternal/Cache/EmbraceUploadCache.swift
@@ -113,7 +113,7 @@ class EmbraceUploadCache {
                 if let uploadData = try context.fetch(request).first {
                     uploadData.data = data
                     uploadData.payloadTypes = payloadTypes
-                    try context.save()
+                    coreData.save()
                     return true
                 }
             } catch {
@@ -133,13 +133,12 @@ class EmbraceUploadCache {
                 attemptCount: 0,
                 date: Date()
             ) {
-
-                do {
-                    try context.save()
-                    return true
-                } catch {
-                    context.delete(record)
+                coreData.performOperation { _ in
+                    if !coreData.saveIfNeeded() {
+                        context.delete(record)
+                    }
                 }
+                return true
             }
             return false
         }
@@ -165,7 +164,7 @@ class EmbraceUploadCache {
                     context.delete(uploadData)
                 }
 
-                try context.save()
+                coreData.save()
             }
         } catch {
             logger.error("error checking count limit:\n\(error.localizedDescription)")
@@ -202,11 +201,7 @@ class EmbraceUploadCache {
 
             uploadData.attemptCount = attemptCount
 
-            do {
-                try self?.coreData.context.save()
-            } catch {
-                self?.logger.warning("Error upading attempt count:\n\(error.localizedDescription)")
-            }
+            self?.coreData.save()
         }
     }
